src/requirements.txt
fastapi==0.115.12
grpcio==1.70.0
pandas==2.2.3
protobuf==6.30.2
pydantic==2.11.3
scikit_learn==1.6.1
scikit_surprise==1.1.4
scipy==1.15.2
